+++ src/components/carousel/Carousel.js
@@ -1,15 +1,15 @@
 import React from "react";
 import PropTypes from "prop-types";
-import "./style.scss";
+import "./carousel.scss";
 
 function Carousel({ image, title }) {
   return (
     <div className="carousel">
       <div className="overlay">
         <h2>{title}</h2>
-        <div className='btn btn--animated'>Play Trailer</div>
+        <div className="btn btn--animated">Play Trailer</div>
       </div>
-      <img src={image} alt={title}/>
+      <img src={image} alt={title} />
     </div>
   );
 }
