+++ src/components/carousel/carousel.scss
