+++ src/components/contact-info/index.js
@@ -1,0 +1,1 @@
+export { default } from "./contact-info";
