+++ src/components/footer/Footer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./style.scss";
+import "./footer.scss";
 
 function Footer() {
   return (
