+++ src/components/footer/index.js
@@ -1,0 +1,1 @@
+export { default } from "./footer";
