+++ src/components/movie-card/index.js
@@ -1,0 +1,1 @@
+export { default } from "./movie-card";
