+++ src/components/movie-card/movie-card.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
-import "./style.scss";
+import "./movie-card.scss";
 
 const MovieCard = ({ movie }) => (
   <div
