+++ src/components/movie-card/movie-card.scss
