+++ src/components/movie-list-card/index.js
@@ -1,0 +1,1 @@
+export { default } from "./movie-list-card";
