+++ src/components/movie-list-card/movie-list-card.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./style.scss";
+import "./movie-list-card.scss";
 
 function MovieListCard({ movie }) {
   return (
