+++ src/components/movie-list-card/movie-list-card.scss
