+++ src/components/show-time-tab/show-time-tab.scss
