+++ src/pages/movie-details/index.js
@@ -1,0 +1,1 @@
+export { default } from "./movie-details";
