+++ src/pages/movie-details/movie-details.scss
